feat: revised routes
