+++ app.js
@@ -12,7 +12,7 @@ const methodOverride = require('method-override')
 const passport = require('./config/passport')
 
 const handlebarsHelpers = require('./helpers/handlerbars-helpers')
-const { pages, apis } = require('./routes')
+const router = require('./routes')
 
 const app = express()
 const port = process.env.PORT || 3000
@@ -43,8 +43,7 @@ app.set('view engine', 'hbs')
 // use helpers.getUser(req) to replace req.user
 // use helpers.ensureAuthenticated(req) to replace req.isAuthenticated()
 
-app.use('/api', apis)
-app.use(pages)
+app.use(router)
 
 app.listen(port, () => {
   console.info(`Example app listening on http://localhost:${port}`)
