+++ routes/modules/users.js
@@ -1,7 +1,7 @@
 const express = require('express')
 const router = express.Router()
 
-const userController = require('../../../controllers/apis/user-controller')
+const userController = require('../../controllers/user-controller')
 
 router.get('/:id/followings', userController.getUserFollowings)
 router.get('/:id/followers', userController.getUserFollowers)
